+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix in-game mod list tooltip position on other resolutions
 * Allow multiple Prefix, Postfix, and Insert patches to exist in a single patch class
   * Use the SpirePrefixPatch, SpirePostfixPatch, and SpireInsertPatch annotations to mark methods
+  * If using a locator, Insert must specify locator with the `locator` parameter of SpireInsertPatch
 * Allow Class types to be used in SpirePatch
   * No longer have to type the fully qualified class name